requirements.txt
# Core dependencies
torch>=1.12.0
numpy>=1.21.0
pandas>=1.3.0
scanpy>=1.9.0
anndata>=0.8.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning
pytorch-lightning>=1.7.0  # optional, for cleaner training

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0
umap-learn>=0.5.0

# Data processing
h5py>=3.7.0
tables>=3.7.0
loompy>=3.0.6

# Network analysis
networkx>=2.8.0

# Progress bars and utilities
tqdm>=4.64.0
pyyaml>=6.0

# Jupyter support (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=6.4.0

# Testing (optional)
pytest>=7.1.0
pytest-cov>=3.0.0

# Documentation (optional)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Code quality (optional)
black>=22.0.0
flake8>=4.0.0
mypy>=0.960